flask-demo/requirements.txt
Flask==1.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
keras-tuner==1.0.2
numpy==1.18.0
opencv-python==4.4.0.46
opencv-python-headless==4.5.1.48
pandas==1.1.3
Pillow==7.2.0
Werkzeug==1.0.1
matplotlib3.3.2
Flask-Caching1.10.1
opencv-python4.4.0.46
opencv-python-headless4.5.1.48
scikit-image0.17.2
scikit-learn0.23.2
scikit-optimize0.8.1
scipy1.4.1

